rename: superpowers -> root-actions
