+++ pages/root-actions.js
@@ -5,10 +5,10 @@ import styled, { css } from 'styled-components';
 import AuthenticatedPage from '../components/AuthenticatedPage';
 import Container from '../components/Container';
 import { Box, Grid } from '../components/Grid';
+import ClearCacheForAccountForm from '../components/root-actions/ClearCacheForAccountForm';
+import MergeAccountsForm from '../components/root-actions/MergeAccountsForm';
 import StyledCard from '../components/StyledCard';
 import StyledHr from '../components/StyledHr';
-import ClearCacheForAccountForm from '../components/superpowers/ClearCacheForAccountForm';
-import MergeAccountsForm from '../components/superpowers/MergeAccountsForm';
 import { H1, H2, H3, P, Span } from '../components/Text';
 
 const GRID_TEMPLATE_COLUMNS = ['minmax(220px, 1fr) 6fr'];
